+++ packages/core/src/constants.ts
@@ -177,3 +177,5 @@ export enum PostMessageMethods {
 }
 
 export const SUPPORTED_IMAGE_FORMATS = ['jpg', 'png', 'webp', 'gif', 'avif'] as const;
+
+export const PATTERN_PROPERTY_DIVIDER = '-----';