+++ packages/experience-builder-sdk/src/core/preview/assemblyUtils.spec.ts
@@ -1,11 +1,15 @@
 import type { Entry } from 'contentful';
+import md5 from 'md5';
 import { experienceEntry } from '../../../test/__fixtures__/composition';
 import {
   assemblyGeneratedDesignVariableName,
   createAssemblyEntry,
 } from '../../../test/__fixtures__/assembly';
 import { assets, entries } from '../../../test/__fixtures__/entities';
-import { CONTENTFUL_COMPONENTS } from '@contentful/experiences-core/constants';
+import {
+  CONTENTFUL_COMPONENTS,
+  PATTERN_PROPERTY_DIVIDER,
+} from '@contentful/experiences-core/constants';
 import type { ComponentTreeNode } from '@contentful/experiences-core/types';
 import { EntityStore } from '@contentful/experiences-core';
 import { resolveAssembly } from './assemblyUtils';
@@ -23,7 +27,12 @@ describe('resolveAssembly', () => {
       locale: 'en-US',
     });
 
-    const result = resolveAssembly({ node: containerNode, entityStore });
+    const result = resolveAssembly({
+      node: containerNode,
+      entityStore,
+      parentPatternProperties: {},
+      patternNodeIdsChain: '',
+    });
 
     expect(result).toBe(containerNode);
   });
@@ -37,7 +46,12 @@ describe('resolveAssembly', () => {
 
     const entityStore = {} as unknown as EntityStore;
 
-    const result = resolveAssembly({ node: containerNode, entityStore });
+    const result = resolveAssembly({
+      node: containerNode,
+      entityStore,
+      parentPatternProperties: {},
+      patternNodeIdsChain: '',
+    });
 
     expect(result).toBe(containerNode);
   });
@@ -55,7 +69,12 @@ describe('resolveAssembly', () => {
       locale: 'en-US',
     });
 
-    const result = resolveAssembly({ node: assemblyNode, entityStore });
+    const result = resolveAssembly({
+      node: assemblyNode,
+      entityStore,
+      parentPatternProperties: {},
+      patternNodeIdsChain: '',
+    });
 
     expect(result).toBe(assemblyNode);
   });
@@ -85,7 +104,12 @@ describe('resolveAssembly', () => {
         children: [],
       };
 
-      const result = resolveAssembly({ node: assemblyNode, entityStore });
+      const result = resolveAssembly({
+        node: assemblyNode,
+        entityStore,
+        parentPatternProperties: {},
+        patternNodeIdsChain: '',
+      });
 
       expect(result.children).toHaveLength(1);
       expect(result.children[0].children).toHaveLength(1);
@@ -96,6 +120,51 @@ describe('resolveAssembly', () => {
       expect(entityStore.unboundValues).toEqual({
         ...experienceEntry.fields.unboundValues,
         ...assemblyEntry.fields.unboundValues,
+      });
+    });
+
+    it('should return an assembly node with parent patternProperties', () => {
+      const patternPropertyId = md5('assembly-id') + PATTERN_PROPERTY_DIVIDER + 'patternPropertyId';
+      const patternPropertyId2 =
+        md5('assembly-id') + PATTERN_PROPERTY_DIVIDER + 'patternPropertyId2';
+      const assemblyNode: ComponentTreeNode = {
+        definitionId: 'assembly-id',
+        id: 'assembly-id',
+        variables: {},
+        children: [],
+        patternProperties: {
+          [patternPropertyId]: {
+            contentType: 'testContentType',
+            path: '/1230948',
+            type: 'BoundValue',
+          },
+        },
+      };
+
+      const result = resolveAssembly({
+        node: assemblyNode,
+        entityStore,
+        patternNodeIdsChain: 'assembly-id',
+        parentPatternProperties: {
+          [patternPropertyId2]: {
+            contentType: 'testContentType',
+            path: '/4091203i9',
+            type: 'BoundValue',
+          },
+        },
+      });
+
+      expect(result.patternProperties).toEqual({
+        ['patternPropertyId']: {
+          contentType: 'testContentType',
+          path: '/1230948',
+          type: 'BoundValue',
+        },
+        ['patternPropertyId2']: {
+          contentType: 'testContentType',
+          path: '/4091203i9',
+          type: 'BoundValue',
+        },
       });
     });
 
@@ -111,7 +180,12 @@ describe('resolveAssembly', () => {
         children: [],
       };
 
-      const result = resolveAssembly({ node: assemblyNode, entityStore });
+      const result = resolveAssembly({
+        node: assemblyNode,
+        entityStore,
+        parentPatternProperties: {},
+        patternNodeIdsChain: '',
+      });
 
       expect(result.children[0].variables.cfWidth).toEqual({
         type: 'DesignValue',
@@ -126,7 +200,12 @@ describe('resolveAssembly', () => {
         children: [],
       };
 
-      const result = resolveAssembly({ node: assemblyNode, entityStore });
+      const result = resolveAssembly({
+        node: assemblyNode,
+        entityStore,
+        parentPatternProperties: {},
+        patternNodeIdsChain: '',
+      });
 
       expect(result.children[0].variables.cfWidth).toEqual({
         type: 'DesignValue',