+++ packages/experience-builder-sdk/src/core/preview/assemblyUtils.ts
@@ -1,4 +1,5 @@
 import { checkIsAssemblyNode, EntityStore } from '@contentful/experiences-core';
+import md5 from 'md5';
 import type {
   ComponentPropertyValue,
   ComponentTreeNode,
@@ -7,6 +8,7 @@ import type {
   PatternProperty,
 } from '@contentful/experiences-core/types';
 import { resolvePrebindingPath, shouldUsePrebinding } from '../../utils/prebindingUtils';
+import { PATTERN_PROPERTY_DIVIDER } from '@contentful/experiences-core/constants';
 
 /** While unfolding the assembly definition on the instance, this function will replace all
  * ComponentValue in the definitions tree with the actual value on the instance. */
@@ -37,20 +39,17 @@ export const deserializeAssemblyNode = ({
         patternProperties,
         variable: instanceProperty,
       });
+      const path = resolvePrebindingPath({
+        componentSettings,
+        componentValueKey,
+        patternProperties,
+      });
 
-      if (usePrebinding) {
-        const path = resolvePrebindingPath({
-          componentSettings,
-          componentValueKey,
-          patternProperties,
-        });
-
-        if (path) {
-          variables[variableName] = {
-            type: 'BoundValue',
-            path,
-          };
-        }
+      if (usePrebinding && path) {
+        variables[variableName] = {
+          type: 'BoundValue',
+          path,
+        };
 
         // For assembly, we look up the variable in the assembly instance and
         // replace the ComponentValue with that one.
@@ -59,6 +58,8 @@ export const deserializeAssemblyNode = ({
           type: 'UnboundValue',
           key: instanceProperty.key,
         };
+      } else if (instanceProperty?.type === 'NoValue') {
+        variables[variableName] = instanceProperty;
       } else if (instanceProperty?.type === 'BoundValue') {
         variables[variableName] = {
           type: 'BoundValue',
@@ -102,15 +103,20 @@ export const deserializeAssemblyNode = ({
     children,
     slotId: node.slotId,
     displayName: node.displayName,
+    patternProperties: node.patternProperties,
   };
 };
 
 export const resolveAssembly = ({
   node,
+  parentPatternProperties,
+  patternNodeIdsChain,
   entityStore,
 }: {
   node: ComponentTreeNode;
   entityStore: EntityStore;
+  parentPatternProperties: Record<string, PatternProperty>;
+  patternNodeIdsChain: string;
 }) => {
   const isAssembly = checkIsAssemblyNode({
     componentId: node.definitionId,
@@ -130,6 +136,31 @@ export const resolveAssembly = ({
     return node;
   }
 
+  const patternProperties: Record<string, PatternProperty> = {};
+
+  const allPatternProperties = {
+    ...parentPatternProperties,
+    ...(node.patternProperties || {}),
+  };
+
+  for (const [patternPropertyKey, patternProperty] of Object.entries(allPatternProperties)) {
+    /**
+     * Bubbled up pattern properties are a concatenation of the node id
+     * and the pattern property definition id. We need to split them so
+     * that the node only uses the pattern property definition id.
+     */
+    const [hashKey, patternPropertyDefinitionId] =
+      patternPropertyKey.split(PATTERN_PROPERTY_DIVIDER);
+
+    const hashedNodeChain = md5(patternNodeIdsChain || '');
+
+    const isMatchingNode = hashKey === hashedNodeChain;
+
+    if (!isMatchingNode) continue;
+
+    patternProperties[patternPropertyDefinitionId] = patternProperty;
+  }
+
   const componentFields = assembly.fields;
 
   const deserializedNode = deserializeAssemblyNode({
@@ -138,10 +169,11 @@ export const resolveAssembly = ({
       id: node.id,
       variables: node.variables,
       children: componentFields.componentTree.children,
+      patternProperties,
     },
     componentInstanceVariables: node.variables,
     componentSettings: componentFields.componentSettings!,
-    patternProperties: node.patternProperties || {},
+    patternProperties,
   });
 
   entityStore.addAssemblyUnboundValues(componentFields.unboundValues);