+++ packages/test-apps/react-vite/src/components/ComponentUsingReferences/index.ts
@@ -1,0 +1,1 @@
+export { ComponentUsingReferences } from './ComponentUsingReferences';